Fix all clippy warnings, enforce formatting, harden CI
- Resolve all clippy lints across workspace (engine, core, bench, test-framework)
- Fix FFI unsafe pointer lint with crate-level allow
- Fix missing test-only imports (SystemTime, Mutex, ReplicationTaskType, TaskKind, HashMap)
- Fix loop-that-never-loops in matching_service (proper condvar wait loop)
- Fix approx_constant errors (3.14 -> 1.5 in tests)
- Fix absurd_extreme_comparisons (unsigned >= 0)
- Fix field_reassign_with_default patterns
- Fix while_let_loop, redundant_pattern_matching, identity_op
- Fix inconsistent_digit_grouping, len_zero, unit_arg
- Fix unused imports/variables across test files
- Fix let_underscore_future in test-framework Drop impl
- Apply cargo fmt across entire workspace
- Remove continue-on-error from CI clippy and dotnet test steps
- Remove pipe-to-tail from clippy CI step (was masking exit code)

+++ velocity-workflow-engine/src/chaos_endurance.rs
@@ -195,7 +195,7 @@ pub fn run_soak_test(config: &SoakTestConfig) -> Arc<SoakTestMetrics> {
 
                 // Optionally send signals
                 if config.enable_signals && rand_val > 0.5 {
-                    let signal_id = (rng_state >> 16) as u64;
+                    let signal_id = rng_state >> 16;
                     engine.signal_workflow(workflow_key, signal_id, vec![1, 2, 3]);
                     metrics.signals_sent.fetch_add(1, Ordering::Relaxed);
                 }
